backend/api/store/serializers.py
from models import Categories
from rest_framework import serializers


class CategorySerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Categories
        fields = ["category_id", "category_name"]
