backend/api/requirements.txt
Django==5.1.7
psycopg2-binary==2.9.10
djangorestframework==3.15.2 
django-filter==25.1